Issue-257: Moved UIPaginated to core module
