+++ app/src/main/java/com/kirchhoff/movies/mapper/person/main/IPersonsMapper.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.mapper.person.main
 
-import com.kirchhoff.movies.data.ui.core.UIPaginated
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.main.NetworkPerson
