+++ app/src/main/java/com/kirchhoff/movies/mapper/review/IReviewListMapper.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.mapper.review
 
-import com.kirchhoff.movies.data.ui.core.UIPaginated
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.details.review.UIReview
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.details.review.NetworkReview
