+++ app/src/main/java/com/kirchhoff/movies/repository/person/PersonsRepository.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.repository.person
 
-import com.kirchhoff.movies.data.ui.core.UIPaginated
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.details.person.UIPersonCredits
 import com.kirchhoff.movies.data.ui.details.person.UIPersonDetails
 import com.kirchhoff.movies.data.ui.main.UIPerson
