+++ app/src/main/java/com/kirchhoff/movies/repository/tv/ITvRepository.kt
@@ -1,7 +1,7 @@
 package com.kirchhoff.movies.repository.tv
 
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
-import com.kirchhoff.movies.data.ui.core.UIPaginated
 import com.kirchhoff.movies.data.ui.details.review.UIReview
 import com.kirchhoff.movies.data.ui.details.tv.UITvDetails
 import com.kirchhoff.movies.data.ui.main.UITv
