+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/PersonsVM.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.ui.screens.main.persons
 
-import com.kirchhoff.movies.data.ui.core.UIPaginated
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.repository.person.IPersonsRepository
 import com.kirchhoff.movies.ui.screens.core.PaginatedScreenVM
