+++ app/src/main/java/com/kirchhoff/movies/ui/screens/similar/movie/SimilarMoviesVM.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.ui.screens.similar.movie
 
-import com.kirchhoff.movies.data.ui.core.UIPaginated
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.repository.movie.IMovieRepository
 import com.kirchhoff.movies.ui.screens.core.PaginatedScreenVM
