+++ app/src/main/java/com/kirchhoff/movies/ui/screens/similar/tv/SimilarTvsVM.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.ui.screens.similar.tv
 
-import com.kirchhoff.movies.data.ui.core.UIPaginated
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UITv
 import com.kirchhoff.movies.repository.tv.ITvRepository
 import com.kirchhoff.movies.ui.screens.core.PaginatedScreenVM
