+++ core/src/main/java/com/kirchhoff/movies/core/ui/paginated/UIPaginated.kt
@@ -1,3 +1,3 @@
-package com.kirchhoff.movies.data.ui.core
+package com.kirchhoff.movies.core.ui.paginated
 
 data class UIPaginated<T>(val page: Int, val results: List<T>, val totalPages: Int)
